moved runPython.ts to lib folder
